requirements.txt
python-dotenv==0.21.1
Telethon==1.28.5
